+++ tests/LatinLetterTest.php
@@ -14,8 +14,6 @@ declare(strict_types=1);
 namespace BakameTest\Period\Visualizer;
 
 use Bakame\Period\Visualizer\LatinLetter;
-use League\Period\Period;
-use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -26,45 +24,42 @@ final class LatinLetterTest extends TestCase
     /**
      * @dataProvider providerLetter
      */
-    public function testGetLabels(Sequence $sequence, string $letter, array $expected): void
+    public function testGetLabels(int $nbLabels, string $letter, array $expected): void
     {
         $generator = new LatinLetter($letter);
-        self::assertSame($expected, $generator->generate($sequence));
+        self::assertSame($expected, $generator->generate($nbLabels));
     }
 
     public function providerLetter(): iterable
     {
         return [
             'empty labels' => [
-                'sequence' => new Sequence(),
+                'nbLabels' => 0,
                 'letter' => 'i',
                 'expected' => [],
             ],
             'labels starts at i' => [
-                'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
+                'nbLabels' => 1,
                 'letter' => 'i',
                 'expected' => ['i'],
             ],
             'labels starts ends at ab' => [
-                'sequence' => new Sequence(
-                    new Period('2018-01-01', '2018-02-01'),
-                    new Period('2018-02-01', '2018-03-01')
-                ),
+                'nbLabels' => 2,
                 'letter' => 'aa',
                 'expected' => ['aa', 'ab'],
             ],
             'labels starts at 0 (1)' => [
-                'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
+                'nbLabels' => 1,
                 'letter' => '        ',
                 'expected' => ['0'],
             ],
             'labels starts at 0 (2)' => [
-                'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
+                'nbLabels' => 1,
                 'letter' => '',
                 'expected' => ['0'],
             ],
             'labels with an integer' => [
-                'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
+                'nbLabels' => 1,
                 'letter' => '1',
                 'expected' => ['A'],
             ],